pyproject.toml
[project]
name = "data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.6",
    "apache-airflow-providers-common-sql>=1.30.3",
    "apache-airflow-providers-postgres>=6.5.2",
    "apache-airflow-providers-telegram>=4.9.2",
    "asyncpg>=0.31.0",
    "google-genai>=1.65.0",
    "lxml>=6.0.2",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta==0.4.71b0",
    "psycopg2-binary>=2.9.11",
    "pycoingecko>=3.2.0",
    "python-dotenv>=1.2.1",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "supabase>=2.27.2",
    "vnstock>=3.5.0",
    "websockets>=15.0.1",
    "yfinance>=1.1.0",
]

[dependency-groups]
dev = [
    "freezegun==1.2.2",
    "mypy==1.18.2",
    "pytest==7.4.0",
    "pytest-asyncio==0.21.1",
    "pytest-cov==4.1.0",
    "pytest-mock==3.11.1",
    "ruff==0.14.2",
    "responses==0.23.3",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
files = ["dags", "tests"]
ignore_missing_imports = true
check_untyped_defs = true
